requirements.txt
certifi==2024.7.4
charset-normalizer==2.0.12
idna==3.7
python-dotenv==0.20.0
requests==2.27.1
urllib3==1.26.19
